func: merge the timer functions with the leader functions and add lock check for var apply
